+++ sphaira/source/utils/devoptab_webdav.cpp
@@ -58,7 +58,7 @@ private:
     int devoptab_close(void *fd) override;
     ssize_t devoptab_read(void *fd, char *ptr, size_t len) override;
     ssize_t devoptab_write(void *fd, const char *ptr, size_t len) override;
-    off_t devoptab_seek(void *fd, off_t pos, int dir) override;
+    ssize_t devoptab_seek(void *fd, off_t pos, int dir) override;
     int devoptab_fstat(void *fd, struct stat *st) override;
     int devoptab_unlink(const char *path) override;
     int devoptab_rename(const char *oldName, const char *newName) override;
@@ -188,6 +188,7 @@ int Device::webdav_dirlist(const std::string& path, DirEntries& out) {
             continue;
         }
 
+        // todo: fix requested path still being displayed.
         const auto href = url_decode(href_x.node().text().as_string());
         if (href.empty() || href == requested_path || href == requested_path + '/') {
             continue;
@@ -483,7 +484,7 @@ ssize_t Device::devoptab_write(void *fd, const char *ptr, size_t len) {
     return ret;
 }
 
-off_t Device::devoptab_seek(void *fd, off_t pos, int dir) {
+ssize_t Device::devoptab_seek(void *fd, off_t pos, int dir) {
     auto file = static_cast<File*>(fd);
 
     if (dir == SEEK_CUR) {
